Use generate defaults, and fix a few remaining issues
