+++ public/app/plugins/panel/annolist/module.tsx
@@ -18,19 +18,19 @@ export const plugin = new PanelPlugin<PanelOptions>(AnnoListPanel)
           options: [
             { value: false, label: 'All dashboards' },
             { value: true, label: 'This dashboard' },
-          ] as any, // does not like boolean, but works fine!
+          ],
         },
       })
       .addRadio({
         category: ['Annotation query'],
         path: 'onlyInTimeRange',
         name: 'Time range',
-        defaultValue: false,
+        defaultValue: defaultPanelOptions.onlyInTimeRange,
         settings: {
           options: [
             { value: false, label: 'None' },
             { value: true, label: 'This dashboard' },
-          ] as any, // does not like boolean, but works fine!
+          ],
         },
       })
       .addCustomEditor({
@@ -47,57 +47,58 @@ export const plugin = new PanelPlugin<PanelOptions>(AnnoListPanel)
         category: ['Annotation query'],
         path: 'limit',
         name: 'Limit',
-        defaultValue: 10,
+        defaultValue: defaultPanelOptions.limit,
       })
       .addBooleanSwitch({
         category: ['Display'],
         path: 'showUser',
         name: 'Show user',
-        defaultValue: true,
+        defaultValue: defaultPanelOptions.showUser,
       })
       .addBooleanSwitch({
         category: ['Display'],
         path: 'showTime',
         name: 'Show time',
-        defaultValue: true,
+        defaultValue: defaultPanelOptions.showTime,
       })
       .addBooleanSwitch({
         category: ['Display'],
         path: 'showTags',
         name: 'Show tags',
-        defaultValue: true,
+        defaultValue: defaultPanelOptions.showTags,
       })
       .addRadio({
         category: ['Link behavior'],
         path: 'navigateToPanel',
         name: 'Link target',
-        defaultValue: true,
+        defaultValue: defaultPanelOptions.navigateToPanel,
         settings: {
           options: [
             { value: true, label: 'Panel' },
             { value: false, label: 'Dashboard' },
-          ] as any, // does not like boolean, but works fine!
+          ],
         },
       })
       .addTextInput({
         category: ['Link behavior'],
         path: 'navigateBefore',
         name: 'Time before',
-        defaultValue: '10m',
+        defaultValue: defaultPanelOptions.navigateBefore,
         description: '',
       })
       .addTextInput({
         category: ['Link behavior'],
         path: 'navigateAfter',
         name: 'Time after',
-        defaultValue: '10m',
+        defaultValue: defaultPanelOptions.navigateAfter,
         description: '',
       });
   })
   // TODO, we should support this directly in the plugin infrastructure
-  .setPanelChangeHandler((panel: PanelModel<PanelOptions>, prevPluginId: string, prevOptions: any) => {
+  .setPanelChangeHandler((panel: PanelModel<PanelOptions>, prevPluginId: string, prevOptions: unknown) => {
     if (prevPluginId === 'ryantxu-annolist-panel') {
       return prevOptions as PanelOptions;
     }
+
     return panel.options;
   });
