+++ public/app/plugins/panel/stat/module.tsx
@@ -1,18 +1,11 @@
-import {
-  escapeStringForRegex,
-  FieldOverrideContext,
-  getFieldDisplayName,
-  PanelOptionsEditorBuilder,
-  PanelPlugin,
-  ReducerID,
-  standardEditorsRegistry,
-} from '@grafana/data';
-import { BigValueColorMode, BigValueTextMode, SingleStatBaseOptions, VizOrientation } from '@grafana/schema';
+import { PanelPlugin } from '@grafana/data';
+import { BigValueColorMode, BigValueGraphMode, BigValueJustifyMode, BigValueTextMode } from '@grafana/schema';
 import { commonOptionsBuilder, sharedSingleStatMigrationHandler } from '@grafana/ui';
 
 import { statPanelChangedHandler } from './StatMigrations';
 import { StatPanel } from './StatPanel';
-import { PanelOptions } from './models.gen';
+import { addStandardDataReduceOptions, addOrientationOption } from './common';
+import { defaultPanelOptions, PanelOptions } from './models.gen';
 import { StatSuggestionsSupplier } from './suggestions';
 
 export const plugin = new PanelPlugin<PanelOptions>(StatPanel)
@@ -38,7 +31,7 @@ export const plugin = new PanelPlugin<PanelOptions>(StatPanel)
           { value: BigValueTextMode.None, label: 'None' },
         ],
       },
-      defaultValue: 'auto',
+      defaultValue: defaultPanelOptions.textMode,
     });
 
     builder
@@ -60,23 +53,23 @@ export const plugin = new PanelPlugin<PanelOptions>(StatPanel)
         name: 'Graph mode',
         description: 'Stat panel graph / sparkline mode',
         category: mainCategory,
-        defaultValue: 'area',
+        defaultValue: defaultPanelOptions.graphMode,
         settings: {
           options: [
-            { value: 'none', label: 'None' },
-            { value: 'area', label: 'Area' },
+            { value: BigValueGraphMode.None, label: 'None' },
+            { value: BigValueGraphMode.Area, label: 'Area' },
           ],
         },
       })
       .addRadio({
         path: 'justifyMode',
         name: 'Text alignment',
-        defaultValue: 'auto',
+        defaultValue: defaultPanelOptions.justifyMode,
         category: mainCategory,
         settings: {
           options: [
-            { value: 'auto', label: 'Auto' },
-            { value: 'center', label: 'Center' },
+            { value: BigValueJustifyMode.Auto, label: 'Auto' },
+            { value: BigValueJustifyMode.Center, label: 'Center' },
           ],
         },
       });
@@ -85,101 +78,3 @@ export const plugin = new PanelPlugin<PanelOptions>(StatPanel)
   .setPanelChangeHandler(statPanelChangedHandler)
   .setSuggestionsSupplier(new StatSuggestionsSupplier())
   .setMigrationHandler(sharedSingleStatMigrationHandler);
-
-function addStandardDataReduceOptions<T extends SingleStatBaseOptions>(
-  builder: PanelOptionsEditorBuilder<T>,
-  includeFieldMatcher = true
-) {
-  const valueOptionsCategory = ['Value options'];
-
-  builder.addRadio({
-    path: 'reduceOptions.values',
-    name: 'Show',
-    description: 'Calculate a single value per column or series or show each row',
-    settings: {
-      options: [
-        { value: false, label: 'Calculate' },
-        { value: true, label: 'All values' },
-      ],
-    },
-    category: valueOptionsCategory,
-    defaultValue: false,
-  });
-
-  builder.addNumberInput({
-    path: 'reduceOptions.limit',
-    name: 'Limit',
-    description: 'Max number of rows to display',
-    category: valueOptionsCategory,
-    settings: {
-      placeholder: '25',
-      integer: true,
-      min: 1,
-      max: 5000,
-    },
-    showIf: (options) => options.reduceOptions.values === true,
-  });
-
-  builder.addCustomEditor({
-    id: 'reduceOptions.calcs',
-    path: 'reduceOptions.calcs',
-    name: 'Calculation',
-    description: 'Choose a reducer function / calculation',
-    category: valueOptionsCategory,
-    editor: standardEditorsRegistry.get('stats-picker').editor as any,
-    // TODO: Get ReducerID from generated schema one day?
-    defaultValue: [ReducerID.lastNotNull],
-    // Hides it when all values mode is on
-    showIf: (currentConfig) => currentConfig.reduceOptions.values === false,
-  });
-
-  if (includeFieldMatcher) {
-    builder.addSelect({
-      path: 'reduceOptions.fields',
-      name: 'Fields',
-      description: 'Select the fields that should be included in the panel',
-      category: valueOptionsCategory,
-      settings: {
-        allowCustomValue: true,
-        options: [],
-        getOptions: async (context: FieldOverrideContext) => {
-          const options = [
-            { value: '', label: 'Numeric Fields' },
-            { value: '/.*/', label: 'All Fields' },
-          ];
-          if (context && context.data) {
-            for (const frame of context.data) {
-              for (const field of frame.fields) {
-                const name = getFieldDisplayName(field, frame, context.data);
-                const value = `/^${escapeStringForRegex(name)}$/`;
-                options.push({ value, label: name });
-              }
-            }
-          }
-          return Promise.resolve(options);
-        },
-      },
-      defaultValue: '',
-    });
-  }
-}
-
-function addOrientationOption<T extends SingleStatBaseOptions>(
-  builder: PanelOptionsEditorBuilder<T>,
-  category?: string[]
-) {
-  builder.addRadio({
-    path: 'orientation',
-    name: 'Orientation',
-    description: 'Layout orientation',
-    category,
-    settings: {
-      options: [
-        { value: VizOrientation.Auto, label: 'Auto' },
-        { value: VizOrientation.Horizontal, label: 'Horizontal' },
-        { value: VizOrientation.Vertical, label: 'Vertical' },
-      ],
-    },
-    defaultValue: VizOrientation.Auto,
-  });
-}
